Not only json but also jsonl are now copied

+++ search-index-typesense/copyManualFiles.sh
@@ -14,9 +14,12 @@ search_index_entries_manual_dir="$(pwd)/search-index-typesense/search-index-entr
 search_index_entries_dir="$(pwd)/search-index-typesense/search-index-entries"
 
 # Copy all .json files from the handmade directory to the output directory
-for file in "$search_index_entries_manual_dir"/*.json; do
-    cp "$file" "$search_index_entries_dir"
+for file in "$search_index_entries_manual_dir"/*.{json,jsonl}; do
+    if [[ -f "$file" ]]; then  # this check ensures it's a file, in case no .json or .jsonl files are found
+        cp "$file" "$search_index_entries_dir"
+    fi
 done
+
 
 
 ############## COPY SITEMAP.XML FROM HANDMADE DIR TO AUTOMATED DIR ##############
